requirements.txt
asgiref==3.8.1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.1
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0rc1
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
